ch(add-controllers): add controllers to abstract routes [finishes #166108011]
