+++ server/routes/users.js
@@ -1,25 +1,8 @@
 import express from 'express';
-import usersModel from '../models/users';
+import usersControllers from '../controllers/users';
 
 const router = express.Router();
 
-router.patch('/:userEmail', (req, res) => {
-  const email = req.params.userEmail;
-  const user = usersModel.findUser(email);
-
-  if (!user) {
-    res.status(404).json({
-      status: 404,
-      message: 'No user matching that email',
-    });
-  } else {
-    const verificationStatus = req.body.status;
-    const updatedUser = usersModel.updateVerificationStatus(verificationStatus, user);
-    res.status(200).json({
-      status: 200,
-      data: updatedUser,
-    });
-  }
-});
+router.patch('/:userEmail', usersControllers.verifyStatus);
 
 export default router;
